feat: Add constants.ts to database folder
